requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
docstring_parser==0.16
executing==2.1.0
Faker==33.3.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.13.1
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.1
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.14
langchain-core==0.3.29
langchain-google-vertexai==2.0.9
langchain-groq==0.2.2
langchain-text-splitters==0.3.4
langsmith==0.2.10
lazy_loader==0.4
lightgbm==4.5.0
lime==0.2.0.1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
narwhals==1.20.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
orjson==3.10.13
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.0
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
selenium==4.27.1
setuptools==75.7.0
shap==0.46.0
shapely==2.0.6
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
stack-data==0.6.3
streamlit==1.41.1
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
xgboost==2.1.3
yarl==1.18.3